#  BENCH: add benchmark for somersd