--- a/.gitignore
+++ b/.gitignore
@@ -280,6 +280,7 @@ scipy/stats/biasedurn.cxx
 scipy/stats/biasedurn.pyx
 scipy/stats/_sobol.c
 scipy/stats/_qmc_cy.cxx
+scipy/stats/_hypotests_pythran.cpp
 scipy/version.py
 scipy/special/_exprel.c
 scipy/optimize/_group_columns.c