https://github.com/scipy/scipy/pull/14308#pullrequestreview-693414732
Since `_tau_b` doesn't seem to be a public function, I only added benchmark for `somersd`.

Before Pythran's speedup:
```
[100.00%] ··· ======== ==========
               n_size            
              -------- ----------
                 10     889±10μs 
                100     106±1ms  
              ======== ============
```
After Pythran's speedup(7x-20x faster):
```
[100.00%] ··· ======== ============
               n_size              
              -------- ------------
                 10      125±5μs   
                100     5.37±0.1ms 
              ======== ============
```

cc @rgommers @serge-sans-paille